[update] move the unit test to each file
