+++ src/elfload/elf_header.rs
@@ -118,3 +118,24 @@ impl ElfHeader {
 	}
 }
 
+#[cfg(test)]
+mod tests {
+	use super::super::*;
+
+	#[test]
+	fn elf_header_test() {
+		let loader = match ElfLoader::try_new("./src/example_elf") {
+			Ok(loader) => loader,
+			Err(error) => {
+				panic!("There was a problem opening the file: {:?}", error);
+			}
+		};
+
+		assert!(loader.is_elf());
+		assert_eq!(loader.elf_header.e_type, 2);
+		assert_eq!(loader.elf_header.e_flags, 5);
+		assert_eq!(loader.elf_header.e_version, 1);
+		assert_eq!(loader.elf_header.e_machine, 243);
+	}
+}
+
