+++ src/elfload/program_header.rs
@@ -72,3 +72,22 @@ impl ProgramHeader {
 }
 
 
+
+#[cfg(test)]
+mod tests {
+	use super::super::*;
+
+	#[test]
+	fn program_header_test() {
+		let loader = match ElfLoader::try_new("./src/example_elf") {
+			Ok(loader) => loader,
+			Err(error) => {
+				panic!("There was a problem opening the file: {:?}", error);
+			}
+		};
+
+		assert_eq!(loader.prog_headers[0].p_type, 1);
+		assert_eq!(loader.prog_headers[0].p_flags, 5);
+	}
+}
+
